+++ packages/terra-menu/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 * Fixed
-  * Added logic to validate toggled item on nested menus are retained while navigating back and forth between them.
+  * Fixed issue where toggled item on nested menus was being reset while navigating back and forth between them.
 
 ## 6.69.0 - (June 12, 2023)